+++ UPDATELOG.md
@@ -3,7 +3,10 @@
 ### 🏆 重大改进
 
 - **应用响应速度提升**：采用全新异步处理架构，大幅提升应用响应速度和稳定性
-- **文件操作性能提升**：优化文件读写性能，减少应用等待时间
+
+### ✨ 新增功能
+
+- **Mihomo(Meta) 内核升级至 v1.19.13**
 
 ### 🚀 性能优化
 
@@ -11,6 +14,8 @@
 - 改进服务管理响应性，减少系统服务操作等待时间
 - 提升文件和配置处理性能
 - 优化任务管理和日志记录效率
+- 优化异步内存管理，减少内存占用并提升多任务处理效率
+- 优化启动阶段初始化性能
 
 ### 🐞 修复问题
