+++ package.json
@@ -38,7 +38,7 @@
     "@mui/icons-material": "^7.3.1",
     "@mui/lab": "7.0.0-beta.16",
     "@mui/material": "^7.3.1",
-    "@mui/x-data-grid": "^8.10.2",
+    "@mui/x-data-grid": "^8.11.0",
     "@tauri-apps/api": "2.8.0",
     "@tauri-apps/plugin-clipboard-manager": "^2.3.0",
     "@tauri-apps/plugin-dialog": "^2.3.3",
@@ -51,7 +51,7 @@
     "ahooks": "^3.9.4",
     "axios": "^1.11.0",
     "cli-color": "^2.0.4",
-    "dayjs": "1.11.13",
+    "dayjs": "1.11.16",
     "foxact": "^0.2.49",
     "glob": "^11.0.3",
     "i18next": "^25.4.2",
@@ -65,7 +65,7 @@
     "react-dom": "19.1.1",
     "react-error-boundary": "6.0.0",
     "react-hook-form": "^7.62.0",
-    "react-i18next": "15.7.2",
+    "react-i18next": "15.7.3",
     "react-markdown": "10.1.0",
     "react-monaco-editor": "0.59.0",
     "react-router-dom": "7.8.2",
@@ -82,10 +82,10 @@
     "@tauri-apps/cli": "2.8.3",
     "@types/js-yaml": "^4.0.9",
     "@types/lodash-es": "^4.17.12",
-    "@types/react": "19.1.11",
-    "@types/react-dom": "19.1.8",
+    "@types/react": "19.1.12",
+    "@types/react-dom": "19.1.9",
     "@vitejs/plugin-legacy": "^7.2.1",
-    "@vitejs/plugin-react": "5.0.1",
+    "@vitejs/plugin-react": "5.0.2",
     "adm-zip": "^0.5.16",
     "commander": "^14.0.0",
     "cross-env": "^10.0.0",
