+++ renovate.json
@@ -38,5 +38,5 @@
     }
   ],
   "postUpdateOptions": ["pnpmDedupe"],
-  "ignoreDeps": ["serde_yaml", "sysinfo", "zip", "criterion"]
+  "ignoreDeps": ["criterion"]
 }
