+++ src-tauri/Cargo.toml
@@ -25,10 +25,10 @@ dunce = "1.0.5"
 log4rs = "1.3.0"
 nanoid = "0.4"
 chrono = "0.4.41"
-sysinfo = { version = "=0.35.2", features = ["network", "system"] }
+sysinfo = { version = "0.37.0", features = ["network", "system"] }
 boa_engine = "0.20.0"
 serde_json = "1.0.143"
-serde_yaml = "0.9.34"
+serde_yaml_ng = "0.10.0"
 once_cell = "1.21.3"
 port_scanner = "0.1.5"
 delay_timer = "0.11.6"
@@ -41,7 +41,7 @@ tokio = { version = "1.47.1", features = [
   "sync",
 ] }
 serde = { version = "1.0.219", features = ["derive"] }
-reqwest = { version = "0.12.23", features = ["json", "rustls-tls", "cookies"] }
+reqwest = { version = "0.12.23", features = ["json", "cookies"] }
 regex = "1.11.2"
 sysproxy = { git = "https://github.com/clash-verge-rev/sysproxy-rs" }
 tauri = { version = "2.8.4", features = [
@@ -60,8 +60,8 @@ tauri-plugin-clipboard-manager = "2.3.0"
 tauri-plugin-deep-link = "2.4.2"
 tauri-plugin-devtools = "2.0.1"
 tauri-plugin-window-state = "2.4.0"
-zip = "=4.2.0"
-reqwest_dav = "0.2.1"
+zip = "4.5.0"
+reqwest_dav = "0.2.2"
 aes-gcm = { version = "0.10.3", features = ["std"] }
 base64 = "0.22.1"
 getrandom = "0.3.3"
@@ -73,12 +73,16 @@ hmac = "0.12.1"
 sha2 = "0.10.9"
 hex = "0.4.3"
 scopeguard = "1.2.0"
-kode-bridge = "0.2.1-rc1"
+kode-bridge = "0.2.1-rc2"
 dashmap = "6.1.0"
 tauri-plugin-notification = "2.3.1"
 console-subscriber = { version = "0.4.1", optional = true }
 tokio-stream = "0.1.17"
-isahc = { version = "1.7.2", features = ["parking_lot"] }
+isahc = { version = "1.7.2", default-features = false, features = [
+  "text-decoding",
+  "parking_lot",
+] }
+
 
 [target.'cfg(windows)'.dependencies]
 runas = "=1.2.0"
