+++ src-tauri/src/enhance/merge.rs
@@ -1,5 +1,5 @@
 use super::use_lowercase;
-use serde_yaml::{self, Mapping, Value};
+use serde_yaml_ng::{self, Mapping, Value};
 
 fn deep_merge(a: &mut Value, b: &Value) {
     match (a, b) {
@@ -54,10 +54,10 @@ fn test_merge() -> anyhow::Result<()> {
     script1: test
   ";
 
-    let merge = serde_yaml::from_str::<Mapping>(merge)?;
-    let config = serde_yaml::from_str::<Mapping>(config)?;
+    let merge = serde_yaml_ng::from_str::<Mapping>(merge)?;
+    let config = serde_yaml_ng::from_str::<Mapping>(config)?;
 
-    let _ = serde_yaml::to_string(&use_merge(merge, config))?;
+    let _ = serde_yaml_ng::to_string(&use_merge(merge, config))?;
 
     Ok(())
 }
